% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnValues.R
\docType{data}
\name{ColumnValues}
\alias{ColumnValues}
\title{ColumnValues}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ColumnValues
}
\description{
ColumnValues
}
\keyword{datasets}
