% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relation.R
\docType{data}
\name{CompositeRelation}
\alias{CompositeRelation}
\title{CompositeRelation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CompositeRelation
}
\description{
CompositeRelation
}
\keyword{datasets}
