% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operatorResult.R
\docType{data}
\name{CubeOperatorTableResult}
\alias{CubeOperatorTableResult}
\title{Result}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CubeOperatorTableResult
}
\description{
Result
}
\keyword{datasets}

