% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relation.R
\docType{data}
\name{JoinMeltOperator}
\alias{JoinMeltOperator}
\title{JoinMeltOperator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
JoinMeltOperator
}
\description{
JoinMeltOperator
}
\keyword{datasets}
