% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relation.R
\docType{data}
\name{JoinOperator1}
\alias{JoinOperator1}
\title{JoinOperator}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
JoinOperator1
}
\description{
JoinOperator
}
\keyword{datasets}
