% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relation.R
\docType{data}
\name{PrefixRelation}
\alias{PrefixRelation}
\title{PrefixRelation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PrefixRelation
}
\description{
PrefixRelation
}
\keyword{datasets}

