% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relation.R
\docType{data}
\name{RenameRelation}
\alias{RenameRelation}
\title{RenameRelation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RenameRelation
}
\description{
RenameRelation
}
\keyword{datasets}
