% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relation.R
\docType{data}
\name{SimpleRelation}
\alias{SimpleRelation}
\title{SimpleRelation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SimpleRelation
}
\description{
SimpleRelation
}
\keyword{datasets}

