% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salve_client.R
\docType{data}
\name{SlaveClient}
\alias{SlaveClient}
\title{SlaveClient}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SlaveClient
}
\description{
SlaveClient
}
\keyword{datasets}

