% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube_query.R
\docType{data}
\name{SQLNotEquals}
\alias{SQLNotEquals}
\title{SQLNotEquals}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SQLNotEquals
}
\description{
SQLNotEquals
}
\keyword{datasets}

