% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\docType{data}
\name{TableProperties}
\alias{TableProperties}
\title{ComputedTable}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TableProperties
}
\description{
ComputedTable
}
\keyword{datasets}

