% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtercen.R
\docType{data}
\name{TercenClient}
\alias{TercenClient}
\title{Tercen Client}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TercenClient
}
\description{
Tercen Client
}
\keyword{datasets}
