% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relation.R
\docType{data}
\name{UnionRelation}
\alias{UnionRelation}
\title{UnionRelation}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
UnionRelation
}
\description{
UnionRelation
}
\keyword{datasets}
